chat: remove obsolete repository tracking
Agent Host enablement made the Agents window repository tracker unreachable. Remove the tracker and keep only the cache invalidation helper used after extension-host requests.\n\n(Written by Copilot)\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
